fix: passive circle bounding box radius is now a double

+++ src/main/java/puttingchallenge/model/collisions/PassiveCircleBoundingBox.java
@@ -18,6 +18,6 @@ public interface PassiveCircleBoundingBox {
     /**
      * @return the radius of the bounding box
      */
-    Integer getRadius();
+    double getRadius();
 
 }
